platform/services/installer/pyproject.toml
# Copyright (C) 2022-2025 Intel Corporation
# LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
[project]
name = "platform-installer"
version = "1.0.0"
description = "Installer of the Geti"
requires-python = ">=3.10, <3.11"

dependencies = [
    "click==8.1.7",
    "click-spinner==0.1.10",
    "cryptography==43.0.1",
    "GPUtil==1.4.0",
    "ifaddr==0.2.0",
    "kubernetes==24.2.0",
    "packaging==24.1",
    "parse==1.19.0",
    "psutil==5.9.1",
    "pyinstaller==6.6.0",
    "PyYAML==6.0.1",
    "rich-click==1.4",
    "tenacity==8.2.1",
    "pytz==2022.5",
    "jinja2==3.1.4",
    "geti-k8s-tools==1.0.0",
    "setuptools==75.1.0",
]

[dependency-groups]
dev = [
    "pytest~=8.3",
    "ruff~=0.11",
    "pre-commit~=4.1",
    "mypy~=1.15",
    "pytest-cov==2.11.1",
    "click==8.1.7",
    "pytest-html~=4.0.0",
    "httpx~=0.23",
    "cryptography==43.0.1",
    "pytest-asyncio~=0.19.0",
    "pytest-mock~=3.14",
]

[tool.uv.sources]
geti-k8s-tools = { path = "../../libs/k8s_tools", editable = true }

[tool.ruff]
target-version = "py310"
line-length = 120

exclude = [
    ".venv*",
]

src = ["app"]

[tool.ruff.lint]
select = ["ARG", "E", "F", "I", "N", "UP", "YTT", "ASYNC", "S", "COM", "C4", "FA", "PIE", "PYI", "Q", "RSE", "RET", "SIM",
    "TID", "TC", "PL", "RUF", "C90", "D103", "ANN001", "ANN201", "ANN205", "FAST"]

ignore = ["N801", "N805","N806","N807", "N818", "COM812", "RET503", "SIM108", "SIM105", "PLR2004",
    "RUF010", "TC001",  "RUF012"]

fixable = ["ALL"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]

# Covers any pytest
"*test*.py" = ["E741", "E402", "E501",
    "S",
    "RET",
    "PL",
    "RUF",
    "SIM",
    "D103",
    "ARG",
    "PYI024",
    "ANN"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.pylint]
max-args=7

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
exclude = "^.*/tests/.*\\.py|app/alembic/.*\\.py"
disable_error_code = ["import-untyped"]
check_untyped_defs = true

show_error_codes = true